sparse BinaryHeap_realloc: use cgraph wrappers for allocation, remove error path
The lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.

The transitive callers of this function (the callers of `BinaryHeap_insert`) are
not prepared to handle failure. Additionally creation of a `BinaryHeap` in
`BinaryHeap_new`, prior to the previous commit, did not check for failure and
assumed allocations will succeed. So it seems reasonable to commit to
exit-on-failure behavior here too.