move 'is_centos' to gvtest.py library for testing
There is an upcoming need to detect CentOS within tests/test_regression.py which
thus needs this function somewhere more accessible. This function actually
became unused in 7db8e665cced464c15dfb4bd449d75e0dbce6ac9 but was overlooked and
not removed then. We retain it now that we have an upcoming use for it.

Gitlab: #2356