cgraph gv_realloc: safe guard the behavior of 'realloc(…, 0)'
ISO C23 introduces a new restriction wherein `realloc` with size 0 becomes
undefined behavior:¹

   …if the `size` is zero, the behavior is undefined.

Graphviz is currently compiled with C99 and it will be a long time before it
migrates to C23. But when it does, it is unlikely anyone will note this obscure
standards clause nor think of its relevance to these allocation wrappers.

The alloc.h wrappers generally follow the principle of taking as much mental
load off the caller as possible (e.g. always zeroing new memory). With this
mindset, it seems wise to ensure reallocation calls with 0 size continue to
behave in unsurprising ways in future.

¹ From the current C23 draft,
  https://www.open-std.org/jtc1/sc22/wg14/www/docs/n3054.pdf