common poly_init: Remove unnecessary rounding of width and height
The POINTS macro both converts inches to points and does rounding.

The INCH2PS macro was introduced in
6f1c9013137e5eb8f4ea90080d38dc63bcecce28, which does the same
conversion, but without the rounding. It seem likely that the intent
was to replace the use of POINTS with INCH2PS in all places were node
width and height is dealt with, but this never happened.