circogen: replace node list implementation with generic list
This reduces the amount of code to maintain going forwards. This also replaces
linked-list traversal which is not well suited to modern hardware with linear
array traversal.

Calls to `appendNodelist` with `NULL` as `one` have been replaced with calls to
`nodelist_append`. As `appendNodelist` was being rewritten, this replacement
allows us to avoid handling the `one == NULL` case entirely in this function.

Where a linked-list traversal using `nodelistitem_t` pointers has been swapped
for an index-based traversal of an array, `nodelist_size` has been used in
preference to `sizeNodelist`. This allows less type casting.

Noted inefficiencies that could be improved in future:

  * `realignNodelist` could do a more complicated calculation upfront, in order
    to support a one-pass swap-in-place rotation instead of repeatedly rotating
    one-by-one.

  * `cloneNodelist` could reserve space in the destination list and then
    `memcpy` into it instead of doing a one-by-one append loop.

  * `concatNodelist` could do something similar to `memcpy` instead of
    one-by-one looping.