core plugin picfontname: avoid modifying span font names during font lookup
This logic was written in an odd way that modified the originating font name of
a span during its lookup. This most likely caused no ill effects except for
users rendering a graph to PIC and then reusing the graph for something else.
But it seems like a latent foot gun worth removing.