common emit_xdot: shrink scope of 'pts'
This makes the role of this variable clearer, makes it more obvious to both
humans and the compiler that it is not necessary to retain its values across
loop iterations, and removes its use altogether when not needed. It may not be
necessary to copy the originating points array at all (the polygon, bezier, and
polyline cases). But these cases invoke a possibly-user-supplied rendering
engine which can do anything to the points array, including modify its members.