rtest: port to Pytest
This change allows the rtest test cases to finally be seen as Pytest test cases.
This has a number of benefits:

  1. Test cases are declared in Python. The definitions are syntax-checked and
     mistakes and typos more readily surface.

  2. Pytest, by seeing native test cases, can report more information about test
     case failures.

  3. rtest test cases are no longer constrained to running serially. Pytest can
     be instructed to run them in parallel.

The translation of tests/tests.txt to Python was considered risky and error
prone. So this was done by temporarily modifying the tests/test_rtest.py parsing
logic to output the test definitions as Python code instead of calling `doTest`.
This output was then copy-pasted into tests/test_rtest.py, following by scanning
tests/tests.txt for inline comments and copying these across. We may have to do
something similar in future if we want to port the other .txt files under
tests/.

Gitlab: closes #2197