dotgen makeCompoundEdge: fix: allow 'ltail' with no arrowhead on tail
It is not clear to me why this assertion was expected to pass. An arrow can have
no arrowhead on either or both of the head or tail for any number of reasons
(adjusted margins, small arrowhead size, `ARR_TYPE_NONE` selected, …) all of
which are orthogonal to whether `ltail` is used.

This commit adjusts logic here to anticipate no arrowhead on the tail, in which
case the box intersection calculation (which would be degenerate ) is skipped).

Gitlab: fixes #1879
Reported-by: geostag, David Flater