Pango plugin pango_textlayout: Adjust yoffset_centerline to better align text within nodes
This change makes vertical alignment of text within nodes more similar
to how a browser renders text within a `div` element with a border.

This was tested by visually comparing a `div` with a border, but
without padding with a Graphviz `box` shaped node with `width=0
height=0 and margin=0`. This was done using both the Courier and the
DejaVu fonts of size 14 pt. Firefox version 111.0.1 on Ubuntu 22.04
was used to render both the HTML and the SVG.