GD plugin gd_textlayout: Adjust yoffset_centerline to better align text within nodes
This change makes vertical alignment of text within nodes more similar
to how a browser renders text within a `div` element with a border.

This was tested by visually comparing a `div` with a border, but
without padding, with a Graphviz `box` shaped node with `width=0
height=0 and margin=0`. This was done using both the Courier and the
DejaVu fonts of size 14 pt. Firefox version 111.0.1 on Ubuntu 22.04
was used to render both the HTML and the SVG.

The `descent_per_pt` constant in `text_bbox() in the SVG analyzer is
also adjusted to adapt to this change. The chosen value of 0.257 is a
compromise between the currently ideal value of 0.25 and something
that better fits upcoming changes that will adjust PangoCairo text
layout and rendering, enable it and make it default for CMake builds
when pangoCairo is available. This upcoming change will cause text to
be slightly differently rendered whether PangoCairo is available or
not, because the ratio between font size and actual height is
different with PangoCairo than with the currently used GD plugin and
also varies slightly with font size.

Empirical tests show that the new `descent_per_pt` value minimizes the
error between the estimated bounding box and the actual bounding box
that graphviz uses in the two cases.