neatogen finishEdge: operate in-place on input 'spl'
All callers of this function do not need the array of points they pass in after
their calls. So we can reduce heap pressure and accelerate this function by
operating destructively on its input.

Note that the points in `spl` originate from either `make_polyline` or
`Proutespline`, both of which allocate and then reuse globals for this. This
should probably be refactored in future to have them both return heap-allocated
points arrays that the caller then owns and has the responsibility for
deallocating.