sparse SparseMatrix_import_dot: remove 'edge_label_nodes' parameter
Both callers of this function were passing in null, indicating they did not need
the result of this. Note this commit also removes the `n_edge_label_nodes`
parameter that was used to return the length of the list. One caller was passing
null for this and the other was passing a value it went on to ignore.