neatogen IMDS_given_dim: fix: avoid final scaling if degenerate
When processing a graph with small distances involved, majorization crashes
when trying to index a “block” that contains no nodes. Back tracking this
reveals these degenerate blocks occur because logic that attempts to sort the
nodes of the current level by `<`/`==`/`>` encounters NaN values. The sorting
logic does not anticipate such values that return false for all three preceding
comparison tests. These NaNs are the result of earlier computation that went
awry, and back tracking this reveals these appear much earlier than
majorization.

This change fixes the final source of such NaNs on all supported platforms
except 32-bit MinGW. There appears to be one or more remaining NaN sources on
32-bit MinGW.

When previous computation is degenerate and `uniLength` has been forced to 0,
this final scaling destroys data that has been computed. We can detect and avoid
this case, allowing such graphs to finally be rendered. Whether the result is
useful or not is yet to be seen, but they can at least be processed now.

Gitlab: #1554
Reported-by: dumblob