neatogen standardize: fix: anticipate degenerate length
When processing a graph with small distances involved, majorization crashes
when trying to index a “block” that contains no nodes. Back tracking this
reveals these degenerate blocks occur because logic that attempts to sort the
nodes of the current level by `<`/`==`/`>` encounters NaN values. The sorting
logic does not anticipate such values that return false for all three preceding
comparison tests. These NaNs are the result of earlier computation that went
awry, and back tracking this reveals these appear much earlier than
majorization.

This change fixes one of the sources of such NaNs. With small distances, the
length computation here produces `0`. Attempting to scale by this destroys any
information in the vector, overwriting every point with 0. We can avoid this by
returning before this occurs. This technically no longer does the
standardization it claims, but this seems the right action in this degenerate
case.

Gitlab: #1554