neatogen IMDS_given_dim: fix: anticipate degenerate distance computation
When processing a graph with small distances involved, majorization crashes
when trying to index a “block” that contains no nodes. Back tracking this
reveals these degenerate blocks occur because logic that attempts to sort the
nodes of the current level by `<`/`==`/`>` encounters NaN values. The sorting
logic does not anticipate such values that return false for all three preceding
comparison tests. These NaNs are the result of earlier computation that went
awry, and back tracking this reveals these appear much earlier than
majorization.

This change fixes one of the sources of such NaNs. With small distances, the
Dijkstra computations that round to `int` (`DistType`) can produce `0`. Coercing
these back to `double` and then inverting eventually results in a division of
`inf / inf`. This is avoided by recognizing this scenario and forcing the result
of the length calculation to 0.

Gitlab: #1554