quartz plugin: fix 'mmap' failure confusion on iOS
`mmap` returns not `NULL` on failure but `MAP_FAILED`. So if this `mmap` failed,
`CGBitmapContextCreate` would then be called with `MAP_FAILED` which would crash
internally. Forcing `buffer` to `NULL` beforehand tells `CGBitmapContextCreate`
it needs to allocate internally.

Whether any of the `mmap` code here is relevant anymore is questionable. While
it is still true that iOS has no dedicated swap/page file,¹ it will also not
page out RW memory which is what these `mmap` calls are creating. My
interpretation (I have not tested this on an iOS device) is that these calls
have identical effect on virtual memory occupancy to regular `malloc`. But this
is left to be explored another day.

¹ More nuance explained in
  https://developer.apple.com/library/archive/documentation/Performance/Conceptual/ManagingMemory/Articles/AboutMemory.html