Pango plugin pango_textlayout: remove agxbuf backing buffer
The amount of data printed to this buffer is user-dependent. There seems no
reason to assume it is less than `BUFSIZ` and thus that using a `BUFSIZ` backing
buffer is an optimization. It is simpler to remove this and let the agxbuf
functions use inline storage or a heap buffer as appropriate.

Removing this removes a -Wshadow warning.