CI: add Ubuntu 23.04 jobs
In contrast to the existing Ubuntu CI jobs, the Docker file introduced in this
commit ignores requirements.txt and doc/infosrc/requirements.txt and installs
the one Python package it needs, Pytest, from the native operating system
repositories. Debian and Ubuntu have adopted a policy change that has come into
effect in Ubuntu 23.04 that prevents Pip from installing package requirements
that conflict with system packages.¹ To work around this, we use the native
package manager instead.

The main effect of this is that CI jobs running this image do not have the
precise version of Pytest from requirements.txt. At time of writing, Pytest
7.2.1 is installed. As Graphviz’ use of Pytest is fairly basic and standard,
hopefully this will not cause problems.

An alternative would be to follow the path the policy is intended to push users
towards: creating a Python virtual environment. But this seems onerous to do for
Graphviz, something that is not a Python project, just to get testing machinery
functional.

Gitlab: closes #2382

¹ https://peps.python.org/pep-0668/
  This Github thread explains the actual observed effect and what is going on,
  https://github.com/python/cpython/issues/102134#issuecomment-1445428402.