common emit_begin_edge: clang-format
We do not usually reflow old code. But there is evidence the indentation in this
function is confusing to readers.

This change was done by running `clang-format` on the whole file and then cherry
picking the relevant hunks rather than `--lines=…`. The latter uses some of the
preceding context in the file to somehow infer the function should be indented
with 4 spaces instead of 2 as per the LLVM style. Doing it as per this commit
should lead to less churn if/when the remainder of the file is reformatted.

Github: closes #2387
Reported-by: Vithanco