common: inline LUT alias lists
9409d1dce754e90b461528428152820b419863b9 refactored this code to make the alias
lists arrays instead of pipe-separated strings. The alias lists were outlined
because there was not an obvious way to write them inline, though this was a
compromise making the code harder to read. To look at a font entry in the LUT,
you needed to cross-reference it with the alias list definitions at the top of
the file.

What was overlooked was that we can use C99 compound literals to write the alias
lists inline, regaining the previous readability of having everything about a
LUT entry in one place.