Autotools/CMake: remove expr find-and-replace of stdio with SFIO
The Autotools and CMake build systems were running a coarse-grained
find-and-replace of stdio operations in Bison-generated code to SFIO operations.
The only stdio operations that the Bison-generated code performs is `fprintf` to
`stderr`. In this scenario, stdio and SFIO are equivalent. Thus this
find-and-replace is unnecessary.

This aligns the Autotools and CMake build systems with what the MS Build build
system has been doing ever since the introduction of Bison
(2f116440971450ca961327be706db7f93a81730a).

Gitlab: #1998