expr excomp: support loading some initial text into the pushback buffer
The libexpr input structure (`Exinput_t`) supports a “pushback” buffer for
stashing some data to then be read back out on the next `lex`. You could imagine
a sequence of operations like:

  1. Read some data from `input->fp`
  2. Realize we do not want to handle it yet or we over-read
  3. Stash the excess into the pushback buffer for the next read to find

This could also be used for lookahead/peek, though `Exinput_t` has a separate
field for 1-character lookahead (`Exinput_t.peek`).

The above is discussed in vague hypothetical terms because prior to this commit
the pushback buffer is unused. Its last use was removed in
26d412040d19e29a5ad2908ce44e16cfe9f1d7d1.

The pushback buffer provides a convenient mechanism for inserting arbitrary data
ahead of the file content about to be read. This will be leveraged in an
upcoming commit to simplify the `excomp` caller.

Gitlab: #1873, #1998