expr compile: use an 'agxbuf' for label instead of a stacked SFIO stream
This is a simplification that takes advantage of the ability to address the
`Exinput_t` pushback buffer introduced in the previous commit. Reading from an
in-memory string is significantly more efficient than going through the SFIO
infrastructure (both for construction of this data as well as for reading it
out) and it allows removing the second last instance of SFIO string streams.

Gitlab: #1873, #1998