fix: clamp nslimit-/nslimit1-based iterations to [0, INT_MAX]
`nslimit` and `nslimit1` attributes are doubles with no limits. This meant users
could set negative values or large positive values that would then result in
integer overflow. The iteration loops accounted for this, but the behavior they
had in these cases was likely not what the user intended.

Overflows outside the range `[INT_MIN, INT_MAX]` could be caused in this way.
This is undefined behavior, as described in C99 §6.3.1.4.1:

  When a finite value of real floating type is converted to an integer type
  other than `_Bool`, the fractional part is discarded (i.e., the value is
  truncated toward zero). If the value of the integral part cannot be
  represented by the integer type, the behavior is undefined.

However UBSan does not catch this, so we cannot easily test for it.

Gitlab: #2391