SFIO sfvprintf: hoist '%!' handling
The only calls into this code pass the statically known format string `"%!"`.
Additionally `%!` directives cannot nest.¹

Taking advantage of this, we can handle `%!` upfront. This is a simplification
in that within the main loop we are now always handling a stacked environment.
This will allow further upcoming simplifications in the `sfvprintf` calling
convention, making it more efficient as well as allowing the compiler greater
ability to diagnose misuse.

Gitlab: #1998

¹ Technically SFIO `sfvprintf` supports arbitrary nesting. But `%!` requires
  also passing a C callback for SFIO to invoke to do the actual printing. The
  only current use of SFIO is in GVPR, in which there is no way to define or
  refer to a C callback function.