SFIO sfvprintf: treat incoming format parameters as empty
`sfvprintf` performed some `_Sffmtpos` work to account for positional specifiers
in format strings:

  // in C:
  sfprintf("%2$s%1$s", "abc", "def"); // prints “defabc”

While this was well-intentioned, the only use of `sfvprintf` is via lib/expr
which passes in the initial parameters that are not visible to the end user
writing a GVPR program:

  // in C:
  sfprintf(sp, "%!", &fmt);

As a result, positional specifiers were off by one from what the user would
expect. E.g. `"%2$s%1$s"` in the situation above would appear to refer to
`"def"` and `"abc"` respectively, but would actually refer to `"abc"` and
`&fmt`. If this were to surface in a GVPR program, it would be extremely
confusing as GVPR has no way of dealing with the (C) type `Fmt_t*`. Thankfully
the callback used to print actual data (`prformat`) simply rejects all format
specifiers with a positional component, so any attempt to use `%…$` results in a
readable error message.

In future, perhaps we should remove positional specific support from `sfvprintf`
given it is unusable.

Gitlab: #1998