core plugin xdot_style: abbreviate into “branchless join” style
Broadly speaking, there are two ways to write a delimiter-separated join in C:

  // Technique 1 (flag)
  bool first = true;
  for (size_t i = 0; i < size; ++i) {
    if (!first)
      printf(", ");
    printf("%s", items[i]);
    first = false;
  }

  // Technique 2 (sep)
  const char *separator = "";
  for (size_t i = 0; i < size; ++i) {
    printf("%s%s", separator, items[i]);
    separator = ", ";
  }

The code affected in this commit was written in the first style and this change
alters it to the second style. The second has one trivial advantage (being
slightly shorter) and one non-trivial advantage (avoiding a branch). Though the
relative performance characteristics in this particular piece of code are
irrelevant, the second style should always be preferred as there are some cases
where it makes a dramatic difference.