vt100 plugin: implement a 24-bit color alternative
Controlling colors of terminal characters has a long history and comes in
multiple flavors: monochrome (1-bit), 8 colors (3-bit), 16 colors (4-bit), 88
colors, 256 colors (8-bit), true color (24-bit). 3-bit color is supported more
or less universally and this is what was implemented in the initial addition of
the vt100 plugin. This change implements 24-bit color, capable of covering the
same RGB space as Graphviz. 24-bit color is now supported by many terminal
emulators.¹

3-bit color remains the default for this plugin under the flag `-Tvt100` because
it is a simple mode reliably supported almost everywhere. `-Tvt100-24bit`
selects the higher fidelity mode.

Now that the plugin supports the lowest common baseline and the uppermost mode,
there is little value in implementing the intermediate modes. All are more
complex (require more code) to support than 24-bit color, and terminals that
care about this stuff are even going beyond 24-bit.² The intermediate modes are
now mostly a historical curiosity.

Detecting what color modes the current terminal supports is subtly complex.³ The
plugin makes no attempt to do this kind of discovery, either before or after
this change. It is up to the user to select a mode and the plugin assumes what
the user has told it is correct.

¹ https://github.com/termstandard/colors#truecolor-support-in-output-devices
² https://en.wikipedia.org/wiki/Sixel
  https://sw.kovidgoyal.net/kitty/graphics-protocol/
³ The curious reader is directed to `man terminfo`,
  https://github.com/termstandard/colors, and the Vim, GNU Screen, and Tmux code
  bases.