debian: generate libgv-php5 installation files and handle SWIG ≥ 4.1.0
As of SWIG 4.1.0, its PHP target no longer produces a PHP source wrapper:¹

  2021-05-04: olly
    [PHP] #1982 #1457 https://sourceforge.net/p/swig/bugs/1339/
    SWIG now only use PHP's C API to implement its wrappers, and no
    longer generates PHP code to define classes.  The wrappers should
    be almost entirely compatible with those generated before, but
    faster and without some previously hard-to-fix bugs.

    The main notable difference is SWIG no longer generates a .php
    wrapper at all by default (only if %pragma(php) code=... or
    %pragma(php) include=... are specified in the interface file).
    This also means you need to load the module via extension=...
    in php.ini, rather than letting the dl() in the generated
    .php wrapper load it (but dl() has only worked for command-line
    PHP for some years now).

    *** POTENTIAL INCOMPATIBILITY ***

As a result of this, Debian packaging would fail when attempted with SWIG ≥
4.1.0 due to trying to copy this non-existent file. There appears to be no way
to construct a .install file that works with both new and old SWIG because no
conditional syntax is supported and wildcard matching cannot achieve what we
want. Instead, we work around this in the build system by generating the
.install file. It should now be possible to package Graphviz on a Debian
platform with SWIG ≥ 4.1.0.

Note that it is still unlikely the resulting libgv-php5 package is usable. It
hardcodes PHP 5, a version users and packagers typically no longer have, and
performs PHP 5 steps that silently fail if you do not have it installed.

Gitlab: #2382

¹ https://www.swig.org/Release/CHANGES