CI: remove all macOS jobs
Gitlab’s shared runners have migrated from x86-64 machines to M1 machines. The
Graphviz project failed to meet the deadline for the cutover of this migration.¹
Attempts to begin the migration now have run into cascading failures.² This
commit disables the macOS CI jobs, moving this breakage off the critical path
that has been blocking Graphviz releases for more than a month. Hopefully the
solution to all these problems can be found and this commit can be reverted in
future.

¹ https://gitlab.com/gitlab-com/runner-saas-macos-access-requests/-/issues/3#note_1394341876
² #2411, #2422, #2423, #2424